refactor: widget folder structure
