+++ lotti/lib/widgets/journal/editor_tools.dart
