+++ lotti/lib/widgets/journal/entry_tools.dart
