+++ lotti/lib/widgets/sync/imap_config.dart
@@ -6,7 +6,7 @@ import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:lotti/blocs/sync/config_classes.dart';
 import 'package:lotti/blocs/sync/encryption_cubit.dart';
 import 'package:lotti/theme.dart';
-import 'package:lotti/widgets/buttons.dart';
+import 'package:lotti/widgets/misc/buttons.dart';
 import 'package:lotti/widgets/sync/qr_widget.dart';
 
 class EmailConfigForm extends StatefulWidget {
